version 495

+++ dyld/trace.c
@@ -3,8 +3,6 @@
  *
  * @APPLE_LICENSE_HEADER_START@
  * 
- * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
- * 
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apple Public Source License
  * Version 2.0 (the 'License'). You may not use this file except in
@@ -46,7 +44,7 @@ int trace_type,
 char *name)
 {
     unsigned int i, j, n;
-    unsigned int namelen;
+    int namelen;
     int save_namelen;
     long parms[NUMPARMS];
     char buf[4];
@@ -72,7 +70,7 @@ char *name)
         namelen = strlen(name);
 	/* fprintf(stderr, "name length: %d name %s: \n", namelen, name); */
         
-        if(namelen > sizeof(parms))
+        if(namelen > (int)(sizeof(parms)))
             namelen = sizeof(parms);
 
         /* Remember the length of name. */
@@ -97,7 +95,7 @@ char *name)
              * more 4 byte chunks.
 	     */
             else{
-                for(n = 0; n < namelen; n++)
+                for(n = 0; n < (unsigned int)namelen; n++)
                     buf[n] = *name++;
 		/*
                  * All characters have been collected.  Now fill the rest of
